requirements.txt
pandas
requests
snowflake-connector-python
pytest
pyyaml